split sandbox into multiple files for organization
